better readability in 'Mocking Unbound Methods' example
In [26.6.3.5](https://docs.python.org/3/library/unittest.mock-examples.html#mocking-unbound-methods) standard library paragraph it is an example with `Foo` class with `foo` method, and it's instance is named `foo` too. I propose to rename the method to `bar`, because it wasn't easy to me to understand what exactly second argument in `patch.object(Foo, 'foo', autospec=True)` does: patching an instance of `Foo` class or it's method.